FPST-515: Create Invitations table
We currently create users directly when grant team members are added
and when users try to authenticate via magic links, but this could lead
to a lot of bad data in the user table (users never logging in, incorrect
email addresses, etc.)

This adds an 'Invitations' table to the database to act as a kind of
provisional user table for Deliver Grant Funding users authenticating
with SSO, with the invitation being 'claimed' when a user signs in for
the first time and a User is created in the database.

Also adds a `last_login_utc` column to the User table so that we can
start to track when users log in (via sso or magic link), rather than
having to rely on whether a User's `name` field is populated in the
database to indicate they've signed in with SSO.

Magic links likely won't use the 'Invitation' functionality at all,
at least for now when there's no invitation concept for Access Grant
Funding, so instead we'll just wait to create the User until they've
claimed the magic link and signed in.