FSPT-592: refactor expressions context handling
This tweaks the expression evaluation interface to take a specific
dict-like object instead of an arbitrary dict.

This new 'ExpressionContext' object wraps three levels of dicts:
1. form data
2. submission data
3. expression data

The ExpressionContext is immutable and the wrapped dicts are immutable
too. This means that we can pass them around and use them across
multiple validators without worrying about copying or merging dicts,
which should save a lot of memory churn and - with some familiarity -
make it easier to track/trust that the right data is in the right place.