FIX: Appropriately manage order when deleting entities
Note: all of the methods covered by this fix are in a "temporary"
interface that doesn't have coverage and is intended to make things work
for now. This commit doesn't intend to change that but it should stop us
running into errors if removing and then adding again.

Deleting questions, forms or sections and then trying to add another
currently raises a `DuplicateValueError` for their ordering constraints
(i.e `uq_question_order_form`).

This is especially noticable when you remove the first entity in the
ordered list. When the entity is removed the order of all of the other
entities remains the same (in this example if you have 0, 1, 2 and
remove 0, the two remaining will stay at 1 and 2).

If you try and add a new entity now it will be assigned the order "2" as
there are 2 other entities remaining and the order index starts at 0 -
this will then conflict with the existing entity with the order 2.

If we want to lean on the ORM to manage the ordering using an ordered
list we'll need to tell it that the entity has been removed which will
trigger an appropriate update to the existing elements order and means any
new entity then added will have a unique value based on the max of
whats there.

Steps to reproduce:
- remove the first question on a form
- add a new question to the form