FSPT-515: Change magic link to create user on claim
Previously we would create a User when a magic link was created. This
could lead to bad data in the database. This changes the behaviour
so the User is only created when a magic link is claimed.

Also changes `usable` to `is_usable` as a SQLAlchemy hybrid_property
which can be used in SQL statements and in python code.