FSPT-565: block the ask a question page if the question isn't visible
This is an additional integrity check that should stop the user ever
mistakenly needing to answer a question where the conditions are no
longer met.

This could happen when navigating through the browsers history or
directly copying a link to a question page - we want to ensure that no
data can be overriden or broken-looking pages are shown as conditions
change.