Spike: change interface for expressions
It feels like the core logic for evaluating an expression shouldn't care
so much about some of our domain models - it simply needs a statement to
evaluate and the contextual data to pass in.

This modifies the internal evaluator interface to just take those two
things. It does mean that the external interface (eg `evaluate`) needs
to now be responsible for converting the domain objects into agnostic
data structures, which means that it needs to do the context combining
itself. We don't expect many things to call into the internal layer, so
this shouldn't introduce a large risk of duplication or
doing-the-wrong-thing.

Just thinking through whether it's better to have the core logic in
`_evaluate_expression_with_context` be a bit more domain agnostic and
easier to test. I think we will need the external interfaces (eg
`evaluate`) to still care about the domain so that they can take
responsibility for doing the right thing with merging contexts etc.