Clean up some of the question ID mangling
There are a few places across the codebase where we need to standardise
the question ID into a format that can be used dynamically by Python:

- when attaching question fields to flask forms
- when evaluating expressions

The existing function to do this has spread out into a few corners of
the codebase and started to feel a bit leaky. This makes a small attempt
to clean that up by moving the transformation logic into a mixin, and
adding that to our main two models that need to access the transformed
value: Questions and (pydantic) managed expressions.