Sync removed roles from enum
In https://github.com/communitiesuk/funding-service/pull/330 we removed
a few roles from the role enum, but didn't generate a migration for
them.

This generates that migration using the `alembic_postgresql_enum`
extension, which plugs enum values into alembic for automatic sync
detection. This should make it harder to miss in the future.

When dropping values, we need to first delete (or appropriately handle)
any rows that use the old enum values.