Deserialize answers from submission data
The serialized submission.data blob is a JSON representation of the
answer a user provided. For basic data types like text and integers this
will be equivalent, but when we start thinking of complex datatypes,
we'll likely want to deal with these in python code as more complex
objects directly instead of raw dictionaries.

This updates our expression context to fetch answers through our
deserializer, which will convert the JSON data into Pydantic root
models.