Refactor to remove `from_expression`
@sfount nicely pointed out that if we pass the `expression` into
`build_managed_expression_form` and feed that through to the wtform
field creation, we can pull default values for each field directly in
`build_form_fields`. Then we don't need to have a separate method to
pull the data for those fields and can encapsulate it all in one place.