FIX: explicitly call reorder on the ordering list
Following a very sensible suggestion to pass in the ORM model dirctly
and not separately fetch the parent, there's some unexpected behaviour
around when and where the entities are being removed from the parent
list after calling `db.session.delete()`

As the initial problem we're trying to solve by calling `.remove()` on
the parent list is having it recognise that the order needs updating
this instead opts to try and have a sensible interface and directly tell
the `OrderingList` that it needs to recalculate order after deletion.

This hasn't gotten to the bottom of how we'd expect this to work but
feels like its more predictable and resolves the underlying issue.

Note this accurately reflects to the rest of the code that the lists on
models are `OrderingList`s but it looks like that extension doesn't pass
on its type signature so mypy flags it as an untyped call - I've opted
to ignore that for now as the types suggested here are accurate, we
could maybe patch that upstream in the library.