FSPT-598: bootstrap access grant funding
This bootstraps the main 'access grant funding' part of the codebase
with a blueprint and a base template (based off 'deliver grant
funding').

It also sets up another section of the developers area that is scoped
for access grant funding, and adds a basic 'list all grants' view for
platform admins.