FIX: get supported question types
We're using a defaultdict for the registered data types on the managed
expression registry - this lets us safely access properties of the
supported data types even if they've not explicitly been registered.

This ended up causing a minor regression when showing supported question
types that depended on the "keys" property of the registered data types.
If the defaultdict had been accessed for unsupported data types it would
count them as valid supported types.

This tweaks an accessor to get the registered types to discount falsey values
and updates the get supported questions method to use that.