FSPT-598: Access grant funding 'grant' page
Builds out a basic details page for a grant, from the view of an Access
grant funding user.

This is just a developer draft of the page, aimed at proving some part
of the end-to-end funding journey, so will not look like this when real
users access it.

The page just lists all of the collections created for the grant on the
page, and in the future will let the user create/edit/manage a
submission against any of those collections.

This is a fully-public page; if you aren't logged in, you can still see
the page. You won't see any collections and will instead see a link to
the sign in with a magic link page.