FSPT-596: check conditional dependencies when swapping question order
For now this implements guards to stop two questions from being moved
above/ below other questions that they depend on.

We'll probably want to think through a method of generally checking the
integrity of the order of all questions in a form and any type of
dependency (condition, validation, personalised text).

For now the only managed option which can be configured that depends on
other questions are conditions so lets just start with those.

By being specific about question conditions its easy to provide which
questions are the problem specifically through the UI without coming up
with brand new patterns.

This proposes encapsulating the dependency failure as a generic
exception which should then be able to be passed through to the
template.