Fix anonymous users
Our AuthorisationHelper needs to be able to work with non-logged-in
users. By default, Flask-Login provides an instance of
AnonymousUserMixin in this case. This does not have any of our custom
user attributes, like the `roles` list.

Instead, we can create an empty User and return that. This will mean the
authorisation helper can still do things like poke at its roles, but it
will just have an empty list.