FSPT-596: block managed conditions being added with invalid order
This blocks managed conditions from being added if they depend on
answer to questions that have come before the current question. This
doesn't get into generic dependency checking for conditions, validations
and personalised text.

Note this implements the block in two places, this is because the end
user can actually do something about it when they're selecting a
question but we'll also ideally check the integrity at the point where
we're adding the expression for the user.

If there's a failure when the expression is being added its not handled
as there's nothing the user can directly do about it and it means prior
checks/ the flow is broken - we should hear about it in Sentry.