FSPT-596: block a question being deleted if it has dependencies
This adds a final helper to ensure that a question can't be deleted if
any of the other questions in that form depend on it for their
conditional branching.

Becuase the delete method is in a temporary interface file and the route
itself is a developer route there's not much to hook into in terms of
tests here - happy for the reviewer to reason about if this is
sufficient for now.