BAU: remove interstitial list collections page
The developers pages and tools have been heavily leaned on to let us
build and try out setting up collections. The page was originally
designed to let us try out and built lots of different behaviours but
its currently all collections and watching everyone go through the extra
clicks of clicking "manage" and then getting to what they want to see is
painful.

Move listing collections directly on to the page and remove the extra
clicks.