FSPT-636: Seed on consistent column: user ID
Seeding the faker based on user name was logically flawed, because the
faker then overwrites the username with a new generated name.

This means that the new name will then seed the faker in subsequent
exports, which will force a new name.

Let's seed the faker on a value that doesn't get overwritten: the user
ID.