FSPT-602: adds form pages to Access Grant Funding
This commit hooks up a form runner to Access Grant Funding using the
library and macros shared by Deliver Grant Funding.

This implementation of a collections form is built into the access
funding base template and behaves differently from the Deliver Grant
Funding forms (going back to the list of collections when complete for
the time being) but should hopefully share all of its core business/
state logic with Deliver Grant Funding.

That should allow us to update the form runner and macros and be
confident it will work everywhere we need it to.