FSPT-515: Use Invitations for add user to grant flow
Update the add user to grant team flow to add a row to the Invitation
table if a user doesn't exist rather than create the user directly. If
the user does exist they should still be sent the confirmation email
through Notify but are just given the grant team member role for that
grant.