FSPT-515 Updated sso_get_token flow to use Invitations
Now that "pending" grant team members are added to the Invitation table
rather than having a User object created immediately, the sso_get_token
flow can now check the Invitation table rather than the short-term bodge
that existed to upsert users before we'd even checked if they should have
access.

Now, if they don't already exist or have a valid invite (ie. it's expired
or they just don't have one) then we serve them a 403 before a User object
is even created.

This also includes a small change to update the the send_member_confirmation
email so the link takes the user to the grant that relates to that
specific invitation.