FSPT-602: fixup control flow on runner lib
The control flow for the methods that mutate state on the form runner
lib had ended up convoluted to try and avoid returning a bool to
indicate if the view should continue or not. This results in re-raising
errors and then ignoring them which is a double-negative we should
avoid.

Cleans up some naming and form setters as per review nits too.