Fix spaces between links and full stops
Our prettier config was forcing full stops onto new lines. If the full
stop directly followed a link </a>, then a newline would add a space
between the link text and the full stop - which is ugly.

Switching the HTML whitespace sensitivity to `css` tells prettier to
work out the display style of the blocks, and treat whitespace
sensitively and avoid adding or removing whitespace. This lets us tuck
the full stop directly after the </a> tag and not have it moved onto a
new line by prettier (but also, prettier won't force existing full stops
off of newlines if they're on them).