FSPT-602: move submission confirmation screen over to AGF
This commit removes the submission confirmation screen from deliver
grant funding and moves it to access grant funding.

This is likely more consistent with the current implementation in
deliver grant funding allowing form designers to preview the behaviour
and access grant funding actually allowing stakeholders and actual grant
recipients to see what the journey is like or keep track of their
submission.