Move unit tests structure
We had a bit of a mismatch of directory structures for the unit tests:

Some were laid out with the `unit` dir acting as the `app` directory
itself, and others were laid out with an `app` directory inside the
`unit` dir.

This makes us consistent with the integration folder layout - the `app`
directory is skipped; all tests will be on code in the `app` dir so we
don't need that extra layer in the middle.