FSPT-602: remove restrictions on AGF forms
Tweaks the tests to cover any future sub classes for the form runner lib
and removes the restiction on platform admin for these AGF form pages.

This would cause confusion as what you need to access a magic link
wouldn't line up with what you'd be able to do. A future commit will
consider the AGF user properly in-line with their access to orgs and
grants.