Uupdate managed expressions for checkboxes
With a radio question type, you select one answer. It's very easy to
reason about what it means if that answer is within some set of choices.

With checkboxes you can select multiple answers. What should
ChoicesFromList do in this instance? Would the condition only pass if
all answers are within the selected options, or if there's any overlap
at all? This feels harder to explain to a form designer.

We instead add some distinct managed expressions for checkboxes to
help clarify this:
- A specific (single) answer has been selected by the user
- All of x,y,z answers have been selected by the user