Block removing choices that have been used
If a radio/checkbox answer has been chosen in a submission, then we stop
a form designer from removing the choice from the set of choices.

We'll also need to do this with expressions, which is maybe the more
important thing: choices shouldn't be changing when submissions are
live, but might change a lot while the form designer is building the
form and setting up expressions.

Although when testing the submission you might still have choices
selected. Blocking the removal of the choice then is fine - the form
designer might just need to be directed to delete the offending test
submissions? But how do we then catch and 'release' the selected choice.
Gets a bit complex.