FSPT-669: standard interface for submission answers
We need to use answers that have been submitted in a lot of different
parts of the system. Having a single way of (de)serialising these for
use in different places will give us confidence that we are treating
answers in a consistent way. Using pydantic base models also gives us a
good developer experience/intellisense/code completion/etc.