FSPT-577: Add auth method to session & decorators
As an extra layer of authentication security, we want to make sure that
only SSO users can access Deliver Grant Funding routes and functionality,
and also protect against internal users who should be using SSO
somehow authenticating via magic link. The way this is done is by
injecting a variable into the session in both routes to signify which
route was used to authenticate and our existing decorators now have
additional checks baked in to further protect Deliver Grant Funding.

NB. This required adding the variable to the sessions for test fixtures
which is not as easy as just adding the variable to the session (as
FlaskLogin does in `login_user` but needed a context manager of our
anonymous client).