FSPT-667: Allow managed expressons to choose between validation and conditions
At the moment, any managed expression we create will be available as both a
condition and a validator. This is fine for integer questions, where that makes
sense. But once we add radios, we might allow branching of questions based on
specific choices, but there is no reason to try to 'validate' that a specific
choice was selected. The nature of radios+checkboxes does this already.