FSPT-672: Create data source tables, ready for radios
In the next patch we'll be adding a "Radios" question type. These new Data
Source tables will store the choices that are associated with a radio question.
The Data Source table is a 'container' for all of the choices for a question,
and allows us to apply additional constraints to all of the choices for a single
question (eg all choices in a data source must be unique, and ordering must make
sense).