FSPT-672: update radio choices
This patch makes it possible to update the radio choices for an existing radios
question.

The update logic for radio choices (data source items) is a little complex:

There are three important values for an item:

- id, the PK for the row
- key, an arbitrary identifier for the choice
- label, the human-readable text

For radio questions that have their own data source, the `key` is just a
slugified version of the label. But we can imagine that in, for example, a
platform-level data source for local authorities, the key might be an ONS
organisation code for each local authority.

When updating data source items for a question, we merge any options where the
slugified label matches an existing item. So if an old option was "Option 1" and
you change the text to "option 1", both of these slugify to "option-1", so are
considered the same. This should make it easier to make small
typo/casing/formatting fixes without needing to drop a choice and make a new
one, with a new ID, which might cause trouble when we track references to data
source choices used in expressions elsewhere.