FSPT-673: add AnyOf managed condition for radios
Showing questions conditionally based on the answer of some previous radios
question is an extremely common use-case for form building. Providing a managed
expression for this usecase should make it nice and easy for form builders to
set up conditionally routing through a form.

We update the expressons evaluator here to be based off the 'with compound
types' evaluator, but still trim that down to only allow the 'set' compound
type. This is because our expression for AnyOf uses sets.