FSPT-577: Add separate magic link login required decorator
Now that we've introduced the concept of Access Grant Funding Routes,
our existing `login_required` decorator was causing some non-ideal
behaviour by redirecting unauthenticated users to the SSO sign in page
when they should be redirected to the magic link sign in page.

This adds a new decorator explicitly for Access Grant Funding with the
appropriate redirect for unauthenticated users.