FSPT-686: Add "Yes/No" question
We’ve added support for radio questions to the service, allowing users to define
any list of options. But we expect that, very often, form designers will want to
add a simple “yes” / “no” question type. This is such a common use-case that we
think it’s worth building a specific question type just for this specific type
of radio question.