Add a new email question type
Adds a new question type for email addresses. This is similar to the 'single
line of text', but automatically includes validation for an email address-like
format, which should save form designers time from manually having to set up
this validation.

This ended up being easier than adding a managed expression/validation for
'Email' that could be applied to the single line of text data type.

Updates the developer e2e test to run through an email question as well.