WIP: allow the email question type to check
Currently just a thought experiment trying out how much needs to be
added to hook up a new expression for new question type.

This works both as a validator to ensure the question is being answered
with an email address from a certain domain and as a method of branching
- only showing certain questions if a previous answer is from a given
  domain.

If people think this is interesting then we can make it merge-able -
just one call to make about if a generic "includes" is sufficient or if
we'd want to prove whats needed to set a custom simpleeval function to
ensure the domain part of the email address specifically is correct.