Stop 500s on test submission when task has no questions
The existing code would throw a 500 trying to generate a URL for the first
question in a task on the tasklist page, if the task has no questions in it at
all.

This patch fixes that by detecting "no questions" as part of the tasklist macro
and rendering the text without an href. It also changes the status to "No
questions" in red rather than "Not started" in blue, because it can't be
started. 

This is intended to make it easier for a form designer to understand why the
link isn't there, but does change the pattern slightly. We'd never want to have
a live collection in this state, but I think we can rely on that being handled
by human processes. And I think even if it did slip through, showing an end user
this is better than them seeing an arbitrary 500.