Help form designers test+stop testing
Adds a button on individual 'manage form/task' pages so that form designers can
jump straight into testing an individual form. This is easy while we have no
cross-form interactions; we'll have to think through how we might facilitate
this in a world where there are dependencies _between_ forms.

It also increases the height+visibility of the 'testing banner', and adds a
button to any of the form runner pages so that a form designer can bail out of
the testing journey. The theory is that a form designer is looking to test a
very specific behaviour, and as soon as they've done that they'll want to get
back to tweaking/editing the form configuration itself. We shouldn't force them
to do an awkward multi-back nav, or complete an entire form/submission.