FSPT-684: Enhance radios to autocomplete on >x items
When a form designer creates a radios/"choose one" question, they can enter any
number of options for the radios.

Presenting a form-filler with a list of potentially hundreds of options is a bad
experience. When we go over some number of options - chosen in this case to be
20 - we will instead render a select field and try to enhance it into a GOV.UK
Accessible Autocomplete component, which provides a searchable text field.

https://alphagov.github.io/accessible-autocomplete/

We do this using progressive enhancemene, so if the JS fails for any reason the
user can still use the input as a standard dropdown. Not much better than radios
(probably worse!) - but still technically usable. We expect all of our users to
be on work devices with JS enabled+available.