BAU: Update Magic Link e2e test to work with Notify Team keys
Our magic link e2e test was disabled in dev and test and wouldn't work
locally with a Notify Team key as the test would a) try to send a real
email and b) needed to send to a non-communities/test-communities email
address. This was frustrating, as we don't want to make a habit of
skipping tests e2e tests in our deployed environments or causing issues
with local development.

By sending to our service account with @levellingup, which is in our
Notify guest list, we can get around both these issues and still use
an email address we control and have access to. This will need to be
replaced if the department ever retires the levellingup alias/redirect.