BAU: Fix flash message tuple unpacking
In a couple of places where we called Flask's `get_flashed_messages`
we're expecting the errors to come back as tuples to unpack them with
`type, error` but the errors are only coming back as strings as we're
not passing 'with_categories=True'. As for now we only have one category
and are not using the `type`, we should be able to safely remove this
and just deal with the error contents as a single string to avoid a Flask
error when the flash messages try to render.