Flash message after creating a question
We used to redirect back to the 'manage form(task)' page after adding a
question. Quite often, though, a form designer will want to add conditions or
validation to that same question, which means they then have to find the
question again and click back into it.

This felt annoying, so I changed the redirect to the 'edit question' page. But
that page is very visually similar to the 'add question' page, which can make it
feel like the 'add question' didn't quite work.

This attempts to mitigate some of that confusion by adding a flash message that
directs the user to the next actions:

- go back to the form/task page if you're done
- add conditions/validation to this question if you need it

This isn't perfect - ideally I think we'd want either a very brief 'add
question' page, eg only the question text - or for the add+edit pages to be
identical (allowing you to set up conditions+validation before clicking 'create'
at all). Both of those are technically bigger changes, so this is intended as a
short-term patch to help until the full designs are in.