FSPT-673: Catch data source references and flash error
Now that we have the data source item references table, we can check
for any references when updating a data source and flash an error if the
user tries to delete or change one or more data source items which are
depended on one or more question.

A list of data sources might be referenced in multiple conditions on
multiple questions, so if the form designer tries to delete data source
items which one or more questions depend on now flash a warning
highlighting every dependent question and all the data source items
on which they depend that the form designer tried to delete.