Keep users in answer flow when revealing new questions
Previously, when a user was coming from the check your answers page, we would
take them back to CYA as soon as they provide that one answer.

With this page we keep them in the answer flow if there are any other questions
that still need to be answered. This means that if they change an answer, and
that reveals some more questions that need answering, they immediately have a
chance to provide that information.