Refactor e2e test questions for multiple of same type
Allow us to define all of the questions we want to test in a way that lets us
test the same kind of question multiple times. This will let us easily test the
same question under different conditions, eg rendering options. In this example
we care about testing the radios question with a few options (which render as
radios) and a lot of options (which render as an accessible autocomplete). We
also test the 'none of the above'-style answer.