FSPT-685: Pull data source item choices for expressions from DB
When setting up a managed expression for a question with a data source,
we copy the information about the referenced options into the context of
the expression. We also then copy+hardcode the option keys in the
expression statement.

This dynamically references the data source items through the DB
relationships instead.

Also removes some unnecessary additions around managed expressions.