BAU: Fix data source item dependency exception
When you try to delete data source items on which another question
depended on, we raised an error. However, with interfaces using other
interfaces using other interfaces, the way in which this error was
raised was wrong as it would preserve the original data source items but
not rollback any other changes made to the form, as when the exception
was raised it wasn't in a try/except block or using what would
eventually be a base IntegrityError (but rather raising pre-emptively so
as to have the context needed to build useful error messages).

We might want to take a look at where certain interfaces/which interfaces
try to flush changes and raise exceptions so that this can be consistent
and we don't end up with issues like this happening again.