BAU: Fix auth 500 error
Before the auth work around magic link redirects and the set up of some
initial Access Grant Funding pages, we'd been using an abort(500) for
authenticated users trying to return to the sign-in pages if they
user didn't have an email address with an internal domain. Now we had
user flows that required signing in with magic link (and therefore non-
internal emails) this was outdated and incorrect behaviour. Instead we
now redirect those users to the same place we do as the fallback for
magic link login.

This behaviour also highlighted that Sentry wasn't picking up 500 errors
by default. Adding a logger.error() call as part of our 500 error
handler should mean we get notified of all 500 errors in Sentry going
forward.