FSPT-605: rename question to component
This commit renames the question table to component. This will allow the
table to uniformly manage questions, groups and pages maintaining their
shared order, naming and expression properties while allowing the ORM to
subclass on "type" to provide specific properties and keep the code
upstream specific and readable.

All code outside of the database interface should be able to deal with
well typed "Question" models and therefore this change should change
very little up the upstream code. The only place we need to refer to the
more generic component is in the interface when we're instructing the
ORM to fetch or mutate the table itself.