Allow platform admins to downgrade to grant team members
This is a very quick and easy change that will allow platform admins (ie us) to
quickly and temporarily downgrade our accounts to that of a 'grant team member'.
This should make it easy for eg product managers and designers to view the
service through a "read only"/reduced permissions role.

It doesn't give us everything we might want for permission permutations (ie
someone who's a member of multiple grants) - but it gets us quite a lot for very
little effort.

To get platform admin permissions back, the user just needs to sign out and sign
back in again.