FSPT-605: add group interface
Adds basic group component interfaces to ensure that all of the
schema and migrations connect. These interfaces will be used and
extended by the developer pages to let us add and manage groups on the
form management pages.

Note we make data type optional here as it will be specifically
relevant to the question component - this felt more appropriate than
creating a new no-op/ read-only data type for the time being. Because
we're doing that we add an additional check constraint to make sure that
questions having an appropriate data type is continued to be enforced at
the database level.

Adds type guards for questions and groups to make it clear upstream if
they should be using the `data_type` property - this is done only when
type checking to stop it hooking into alembic migrations.